+++ test/actor_test.cpp
@@ -3,6 +3,7 @@
 #include <map.h>
 #include <engine.h>
 #include <actor_descriptions.h>
+#include <character_type.h>
 
 namespace amarlon {
 
@@ -98,7 +99,7 @@ TEST_F(ActorTest, actorEqual_different_container)
   ActorPtr ca2c = ca1c->clone();
 
   AiDescriptionPtr dsc = std::make_shared<AiDescription>();
-  dsc->type = AiType::GenericAi;
+  dsc->type = (int)AiType::GenericAi;
   ca1c->insertFeature( Ai::create( dsc ));
   c1->add(ca1c);
   c2->add(ca2c);
@@ -142,7 +143,7 @@ TEST_F(ActorTest, actorEqual_different_fighter)
   CharacterDescriptionPtr mobDsc(new CharacterDescription );
   mobDsc->level = 123;
   mobDsc->experience = 666;
-  mobDsc->type = CharacterType::Generic;
+  mobDsc->type = (int)CharacterType::Generic;
 
   a1->insertFeature( Character::create(mobDsc) );
